chore: change structure fix: few bugs
