+++ ethereum-operations/src/common/utils.ts
@@ -1,0 +1,21 @@
+import {Finding, FindingType, TransactionEvent} from "forta-agent";
+
+export function handleEventsOfNotice(txEvent: TransactionEvent, findings: Finding[], eventsOfNotice: any[]) {
+  eventsOfNotice.forEach((eventInfo) => {
+    if (eventInfo.address in txEvent.addresses) {
+      const events = txEvent.filterLog(eventInfo.event, eventInfo.address);
+      events.forEach((event) => {
+        findings.push(
+          Finding.fromObject({
+            name: eventInfo.name,
+            description: eventInfo.description(event.args),
+            alertId: eventInfo.alertId,
+            severity: eventInfo.severity,
+            type: FindingType.Info,
+            metadata: { args: String(event.args) },
+          })
+        );
+      });
+    }
+  });
+}
