+++ ethereum-operations/src/subagents/acl-changes/utils.ts
@@ -1,0 +1,5 @@
+import { ethersProvider } from "../../ethers";
+
+export async function isContract(address: string): Promise<boolean> {
+    return (await ethersProvider.getCode(address)) != "0x";
+}
