+++ ethereum-operations/src/subagents/easy-track/agent-easy-track.ts
@@ -1,25 +1,28 @@
 import {
   ethers,
-  BlockEvent,
   TransactionEvent,
   Finding,
   FindingType,
   FindingSeverity,
 } from "forta-agent";
 
-import { ethersProvider } from "./ethers";
+import { ethersProvider } from "../../ethers";
 
 import {
-  EASY_TRACK_EVENTS_OF_NOTICE,
   EASY_TRACK_ADDRESS,
-  MOTION_CREATED_EVENT,
-  INCREASE_STAKING_LIMIT_ADDRESS,
   NODE_OPERATORS_REGISTRY_ADDRESS,
+} from "../../common/constants";
+
+import {
+  INCREASE_STAKING_LIMIT_ADDRESS,
+  EASY_TRACK_EVENTS_OF_NOTICE,
+  MOTION_CREATED_EVENT,
 } from "./constants";
 
-import INCREASE_STAKING_LIMIT_ABI from "./abi/IncreaseStakingLimit.json";
-import NODE_OPERATORS_REGISTRY_ABI from "./abi/NodeOperatorsRegistry.json";
-import { getMotionLink, getMotionType } from "./utils/tools";
+import INCREASE_STAKING_LIMIT_ABI from "../../abi/IncreaseStakingLimit.json";
+import NODE_OPERATORS_REGISTRY_ABI from "../../abi/NodeOperatorsRegistry.json";
+import { getMotionLink, getMotionType } from "./utils";
+import {handleEventsOfNotice} from "../../common/utils";
 
 export const name = "EasyTrack";
 
@@ -32,34 +35,10 @@ export async function initialize(
 
 export async function handleTransaction(txEvent: TransactionEvent) {
   const findings: Finding[] = [];
-
-  handleEasyTrackEventsOfNotice(txEvent, findings);
+  handleEventsOfNotice(txEvent, findings, EASY_TRACK_EVENTS_OF_NOTICE);
   await handleEasyTrackMotionCreated(txEvent, findings);
 
   return findings;
-}
-
-function handleEasyTrackEventsOfNotice(
-  txEvent: TransactionEvent,
-  findings: Finding[]
-) {
-  EASY_TRACK_EVENTS_OF_NOTICE.forEach((eventInfo) => {
-    if (eventInfo.address in txEvent.addresses) {
-      const events = txEvent.filterLog(eventInfo.event, eventInfo.address);
-      events.forEach((event) => {
-        findings.push(
-          Finding.fromObject({
-            name: eventInfo.name,
-            description: eventInfo.description(event.args),
-            alertId: eventInfo.alertId,
-            severity: eventInfo.severity,
-            type: FindingType.Info,
-            metadata: { args: String(event.args) },
-          })
-        );
-      });
-    }
-  });
 }
 
 async function handleEasyTrackMotionCreated(
@@ -121,3 +100,9 @@ async function handleEasyTrackMotionCreated(
     );
   }
 }
+
+// required for DI to retrieve handlers in the case of direct agent use
+exports.default = {
+  handleTransaction,
+  initialize, // sdk won't provide any arguments to the function
+};
