+++ ethereum-operations/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
     "target": "es5",
+    "sourceMap": true,
     "module": "commonjs",
     "outDir": "dist",
     "strict": true,
